Moved Timer to Views with other stuff.
